(fix)logout bug
