+++ client/components/LandingPage.js
@@ -1,5 +1,4 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
+import React from 'react'
 import { NavLink } from 'react-router-dom'
 
 const LandingPage = () => {
