+++ client/components/SinglePage.js
@@ -8,14 +8,13 @@ class SinglePage extends React.Component {
 		this.props.getOnePage(pageId);
 	}
  render(){
-   console.log('SinglePage Props', this.props)
    return(
      <div>
        {
-         this.props.page &&
+         this.props.myPage &&
            <div>
          <h1>Comic Title</h1>
-    <img src={this.props.page.location}/>
+    <img src={this.props.myPage.location}/>
       </div>
        }
       </div>
@@ -25,7 +24,7 @@ class SinglePage extends React.Component {
 
 const mapState = state =>{
   return{
-    page: state.user.curUser.singlePage
+    myPage: state.page.singlePage
   }
 }
 const mapDispatch = dispatch => {
