+++ client/components/Stickers.js
@@ -47,7 +47,7 @@ class Stickers extends React.Component {
 
 const mapStateToProps= state => {
   return {
-    allStickers: state.page.allStickers
+    allStickers: state.sticker.allStickers
   }
 }
