+++ client/components/auth-form.js
@@ -49,7 +49,7 @@ const mapLogin = state => {
   return {
     name: 'login',
     displayName: 'Login',
-    error: state.user.curUser.error
+    error: state.user.error
   }
 }
 
@@ -57,7 +57,7 @@ const mapSignup = state => {
   return {
     name: 'signup',
     displayName: 'Sign Up',
-    error: state.user.curUser.error
+    error: state.user.error
   }
 }
