+++ client/components/navbar.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
 import {logout} from '../store'
-//import logo from '.././logo.svg'
 
 const Navbar = ({handleClick, isLoggedIn}) => (
   <div>
@@ -12,7 +11,7 @@ const Navbar = ({handleClick, isLoggedIn}) => (
       {isLoggedIn ? (
         <div>
           {/* The navbar will show these links after you log in */}
-          <Link to="/home">Home</Link>
+          <Link to="/home">Home</Link>     
           <a href="#" onClick={handleClick}>
             Logout
           </a>
@@ -34,7 +33,7 @@ const Navbar = ({handleClick, isLoggedIn}) => (
  */
 const mapState = state => {
   return {
-    isLoggedIn: !!state.user.curUser.id
+    isLoggedIn: !!state.user.id
   }
 }
