+++ client/store/index.js
@@ -4,8 +4,9 @@ import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
 import page from './page'
+import sticker from './sticker'
 
-const reducer = combineReducers({user, page})
+const reducer = combineReducers({user, page, sticker})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
@@ -14,3 +15,4 @@ const store = createStore(reducer, middleware)
 export default store
 export * from './user'
 export * from './page'
+export * from './sticker'
